assistant.requirements.txt
fastapi
uvicorn
python-dotenv
azure-cognitiveservices-speech
openai
python-multipart
pydub